Moved hitpoints tool
