+++ Arma3ObjectBuilder/tool_hitpoint/clouds.py
