+++ Arma3ObjectBuilder/utilities/__init__.py
@@ -1,5 +1,4 @@
 # In order of dependency
-from . import clouds
 from . import proxy
 from . import masses
 from . import outliner
